requirements.txt
fastapi
uvicorn
mercadopago
python-dotenv
vercel-kv
qrcode[pil]
